Rename browzineJournals and browzineArticleInContext to simply journals and articles
The default URL has primo/browzine/ already in it, so if we use that, we already identify this URL as being related to browzine.
Also, this entire repo has browzine in the name, so the node server being for browzine articles and journals should be fairly clear - and the README should make that clear.